Refactor CommentsList component
