+++ src/pages/iframes/comments/[dtpId].tsx
@@ -1,14 +1,25 @@
 import { GetServerSideProps, NextPage } from "next";
 import Error from "next/error";
 import * as React from "react";
+import styled from "styled-components";
 
 import { CommentList } from "../../../components/сomment-list";
 import { commentsArePaused } from "../../../shared/helpersForComments";
+import { Comment } from "../../../types";
 
 export interface CommentsIframePageProps {
   dtpId?: number;
-  comments?: unknown[]; // Comment[]
+  comments?: Comment[];
 }
+
+const CommentsHeader = styled.h2`
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  font-size: 26px;
+  line-height: 40px;
+  color: #18334a;
+`;
 
 const CommentsIframePage: NextPage<CommentsIframePageProps> = ({
   dtpId,
@@ -20,7 +31,8 @@ const CommentsIframePage: NextPage<CommentsIframePageProps> = ({
 
   return (
     <div>
-      <CommentList />
+      <CommentsHeader>Комментарии - {comments.length}</CommentsHeader>
+      <CommentList comments={comments} />
       <p>
         {commentsArePaused ? (
           "Добавление новых комментариев приостановлено"
@@ -31,6 +43,23 @@ const CommentsIframePage: NextPage<CommentsIframePageProps> = ({
     </div>
   );
 };
+
+const comments: Comment[] = [
+  {
+    id: 1,
+    text: "информация о верификации данных, если координаты изменены при обработке (координаты отличаются от заявленных ГИБДД, но прошли подтверждение модератором).",
+    user: "Павел Кучерягин",
+    date: new Date().toUTCString(),
+    avatarUrl:
+      "https://robohash.org/6ae852fa3a8b1c79dba3f7dc883c1760?set=set4&bgset=&size=200x200",
+  },
+  {
+    id: 2,
+    text: "Оставленная пользователями дополнительная/уточняющая информация",
+    user: "Anna Kravtz",
+    date: new Date().toUTCString(),
+  },
+];
 
 export const getServerSideProps: GetServerSideProps<
   CommentsIframePageProps
@@ -44,7 +73,7 @@ export const getServerSideProps: GetServerSideProps<
     return {
       props: {
         dtpId,
-        comments: [],
+        comments, // TODO: Replace sample with fetched data
       },
     };
   }
